refactor(frontend): extract shared auth/color-scheme/fetch into one package
Reconcile the diverged copies of useAuth, useColorScheme and apiFetch (editor
vs admin) into a single source and consume it from all three frontends via a
root npm workspace.

New package packages/frontend-shared (@regelrecht/frontend-shared), pure-ESM Vue
consumed as raw source (no @nldd dependency, vue as a peer):
- apiFetch / apiFetchJson / apiFetchText / ApiError (the editor's centralised
  implementation)
- useAuth / ensureAuthReady (the editor's canonical implementation)
- useColorScheme(persistence?) — owns <html data-scheme> application and theme
  validation, with a pluggable persistence adapter (server-backed for the
  editor, localStorage default for admin and lawmaking)

Consumers:
- editor: apiFetch and useAuth become thin re-export shims; useColorScheme wraps
  the shared one with its server-backed useUserSettings adapter (the data-scheme
  watchEffect moved out of useUserSettings into the shared composable)
- admin: conforms to the editor — useColorScheme/useAuth now come from the shared
  package; authedFetch stays admin-local glue (no global 401 guard) and sources
  its redirect from the shared login
- lawmaking: consumes color-scheme only (no auth/API)

Workspace + Docker:
- root package.json (workspaces) + root .npmrc + single root lockfile; the three
  per-frontend lockfiles and .npmrc files are removed
- the three frontend Dockerfiles build via the workspace (npm ci over all
  workspace manifests, then npm run build -w <name>); runtime dist paths and the
  WASM copy target updated; .dockerignore negations keep every workspace
  package.json in context

Also aligns admin @nldd/design-system to ^0.8.63.
