+++ frontend-lawmaking/package.json
@@ -8,6 +8,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@regelrecht/frontend-shared": "*",
     "@nldd/design-system": "^0.8.63",
     "vue": "^3.5.38"
   },
