+++ frontend/package.json
@@ -15,6 +15,7 @@
     "test:e2e": "playwright test"
   },
   "dependencies": {
+    "@regelrecht/frontend-shared": "*",
     "@cucumber/gherkin": "^40.0.0",
     "@cucumber/messages": "^33.0.2",
     "@nldd/design-system": "^0.8.63",
