--- packages/admin/frontend-src/.npmrc
@@ -1,1 +1,0 @@
-registry=https://registry.npmjs.org/
