+++ packages/admin/frontend-src/package.json
@@ -8,7 +8,8 @@
     "build": "vite build"
   },
   "dependencies": {
-    "@nldd/design-system": "^0.8.62",
+    "@regelrecht/frontend-shared": "*",
+    "@nldd/design-system": "^0.8.63",
     "vue": "^3.5.38",
     "vue-router": "^5.1.0"
   },
